More installer tweaks

+++ installer.ps1
@@ -91,53 +91,48 @@ try {
     $PythonVer = (python -V).Split()[-1]
     if ($PythonVer -lt "3.7.5") {
         throw "Python too old"
-    }
-    else {
+    } else {
         Write-Host "Python 3.7+ already installed, moving on..."
         $Python = (Get-Command python).Source
-    }    
-    try {
-        Get-Command bcml | Out-Null
-        Write-Host "BCML is already installed, no need for this."
-        PromptQuit
-    } catch {}
-}
-catch {
-    Write-Host "Looks like you need Python! Downloading Python 3.7.9..."
-    try {
-        (New-Object Net.WebClient).DownloadFile("https://www.python.org/ftp/python/3.7.9/python-3.7.9-embed-amd64.zip", "$env:temp\python.zip")
     }
-    catch {
-        Write-Error "Could not download Python. Please check your Internet connection."
-        PromptQuit
+} catch {
+    if (Test-Path -Path "$home\.python\python.exe") {
+        Write-Host "Python already installed by BCML found. Using that, then."
+        $Python = "$home\.python\python.exe"
+    } else {
+        Write-Host "Looks like you need Python! Downloading Python 3.7.9..."
+        try {
+            (New-Object Net.WebClient).DownloadFile("https://www.python.org/ftp/python/3.7.9/python-3.7.9-embed-amd64.zip", "$env:temp\python.zip")
+        } catch {
+            Write-Error "Could not download Python. Please check your Internet connection."
+            PromptQuit
+        }
+        Write-Host "Downloaded Python. Where would you like it to be installed?"
+        Write-Host "(If you leave this blank, it will default to %USERPROFILE%\.python)"
+        Write-Host "> " -NoNewline
+        $PythonDir = Read-Host
+        if ([String]::IsNullOrWhiteSpace($PythonDir)) {
+            $PythonDir = "$home\.python"
+        }
+        try {
+            Expand-Archive -Path "$env:temp\python.zip" -DestinationPath "$PythonDir" -Force
+        } catch {
+            Write-Error "There was a problem extracting the Python package."
+            PromptQuit
+        }
+        $Python = "$PythonDir\python.exe"
+        Write-Host "Setting up pip..."
+        (New-Object Net.WebClient).DownloadFile("https://bootstrap.pypa.io/get-pip.py", "$PythonDir\get-pip.py") | Out-Null
+        & $Python "$PythonDir\get-pip.py" --disable-pip-version-check --no-warn-script-location | Out-Null
+        Set-Content -Path "$PythonDir\python37._pth" -Value "python37.zip`n.`n.\Lib`n.\Lib\site-packages`n.\DLLs`n"
+        Write-Host "Python is all set!"
     }
-    Write-Host "Downloaded Python. Where would you like it to be installed?"
-    Write-Host "(If you leave this blank, it will default to %USERPROFILE%\.python)"
-    Write-Host "> " -NoNewline
-    $PythonDir = Read-Host
-    if ([String]::IsNullOrWhiteSpace($PythonDir)) {
-        $PythonDir = "$home\.python"
-    }
-    try {
-        Expand-Archive -Path "$env:temp\python.zip" -DestinationPath "$PythonDir" -Force
-    }
-    catch {
-        Write-Error "There was a problem extracting the Python package."
-        PromptQuit
-    }
-    $Python = "$PythonDir\python.exe"
-    Write-Host "Setting up pip..."
-    (New-Object Net.WebClient).DownloadFile("https://bootstrap.pypa.io/get-pip.py", "$PythonDir\get-pip.py") | Out-Null
-    & $Python "$PythonDir\get-pip.py" --disable-pip-version-check --no-warn-script-location | Out-Null
-    Set-Content -Path "$PythonDir\python37._pth" -Value "python37.zip`n.`n.\Lib`n.\Lib\site-packages`n.\DLLs`n"
-    Write-Host "Python is all set!"
 }
 
 try {
     Write-Host "Installing latest BCML from PyPI..."
-    & $Python -m pip install bcml --disable-pip-version-check --no-warn-script-location | Out-Null
-}
-catch {
+    & $Python -m pip install bcml --upgrade --disable-pip-version-check --no-warn-script-location | Out-Null
+} catch {
     Write-Error "BCML did not install successfully from PyPI."
     PromptQuit
 }
@@ -182,14 +177,13 @@ try {
     $UninstallShortcut.Arguments = "$InstallDir\Uninstall.ps1"
     $UninstallShortcut.Description = "Uninstall"
     $UninstallShortcut.Save()
-}
-catch {
+} catch {
     Write-Host "There was a problem creating shortcuts. BCML is still there; you can run it with:"
     Write-Host "$Python -m bcml"
     Write-Host "Feel free to make your own shortcut to this command."
 }
 
-Write-Host "BCML has been installed! Go ahead and launch it from the shortcut if you wish.`n"
+Write-Host "The latest version of BCML has been installed! Go ahead and launch it from the shortcut if you wish.`n"
 Write-Host "Just a reminder: I have a full time job, a Masters program, and 6 children."
 Write-Host "So if you like what I do, consider supporting my Patreon (link above)."
 Write-Host "It's what convinces my wife (partially) this is all worth the time.`n"
